Add shortcut to toggle emacs input
The ideas is to link this script to Super+e so we can toggle the
Gnome 'emacs input' option on the fly. This option should overwrite
current keybindings and apply emacs keybindings. This would be great
editing text in a browser or any non-emacs app. The issue is that this
gnome option does not work great and some emacs keybinding (e.g. Ctrl+b
in Gmail) will not work. I add it any way as a proof of concept.